+++ server/src/auth/auth.module.ts
@@ -4,9 +4,7 @@ import { AuthController } from './auth.controller';
 import { UsersModule } from 'src/users/users.module';
 import { PassportModule } from '@nestjs/passport';
 import { JwtModule } from '@nestjs/jwt';
-import { LocalStrategy } from './strategies/local.strategy';
-import { JwtStrategy } from './strategies/jwt.strategy';
-import { JwtRefreshStrategy } from './strategies/jwt-refresh.strategy';
+import { LocalStrategy, JwtRefreshStrategy, JwtStrategy } from './strategies';
 
 @Module({
   imports: [UsersModule, PassportModule, JwtModule],
