+++ server/src/auth/guards/index.ts
@@ -1,0 +1,3 @@
+export * from './jwt-auth.guard'
+export * from './jwt-refresh-auth.guard';
+export * from './local-auth.guard';
