+++ server/src/auth/utils/jwt.util.ts
@@ -1,0 +1,6 @@
+export function generateJwtToken(payload: object, secret: string, expiresInMs: number): string {
+    return this.jwtService.sign(payload, {
+        secret,
+        expiresIn: `${expiresInMs}ms`,
+    });
+}
